Workaround canvas size issue.
react-three-fiber calls THREE.Renderer.setSize with a value obtained
from `useMeasure`, and from a CSS perspective the HTMLCanvasElement is
still small, even after you zoom in.

This is the library that provides `useMeasure`:
https://github.com/pmndrs/react-use-measure

We can configure this somewhat by adding a `resize` attribute to
`<Canvas>`, but unfortunately there's no "override" feature in the
use-measure library.

One idea is to fork the above library and add an override feature.
Another idea is to set `scene.onBeforeRender`. Inside the callback we
can clobber the resize that was done by react-three-fiber.

This PR is an implementation of the `onBeforeRender` idea, which seems
to work. I think a nicer fix would be to post a PR to one of the
upstream `pmndrs` libraries to allow this behavior to be configured.